analizer/requirements.txt
pika
python-decouple
transformers
tensorflow
tf-keras
emoji
pysentimiento